requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
yookassa==3.0.0
aiohttp==3.9.1

